Initial checkin of low-level pandora api crate
This is a rough outline of a crate for making calls to Pandora web api
functions.  The aim is to provide some basic struct definitions for the
most basic forms of all the calls and responses, and allow clients of
the code to easily define their own custom datatypes tailored to their
use case.

I could not have done this alone - I've built on the shoulders of giants
here, and those shoulders should all be correctly credited in the
source, at least insofar as my own sources are correctly attributed
themselves.